[Performance] Dump autoload to optimize class lookup
Dumps a class array used to lookup classes ala eZ Publish legacy, if class is not found in this default dynamic file system loading is used instead. On PHP 5.6 and higher where the loading of this class array is optimized by placing it in opcache shm memory, thus avoid slow down caused by loading a gigantic autoload array on eZ Platform, speedup is typically around 5-10%.

Was planning to solve this in eZ Platform source code, however QA found issue with that so might have to fall back to instruct in doc to run it, and hence add it here as well.

Adding it here does not hurt no matter outcome in eZ Platofmr source, just some extra seconds spent during build.